+++ graylog2-web-interface/src/views/components/actions/FieldActions.jsx
@@ -6,8 +6,8 @@ import { PluginStore } from 'graylog-web-plugin/plugin';
 import FieldType from 'views/logic/fieldtypes/FieldType';
 import { ActionContext } from 'views/logic/ActionContext';
 import type { QueryId } from 'views/logic/queries/Query';
-import OverlayDropdown from './OverlayDropdown';
-import style from './Field.css';
+import OverlayDropdown from '../OverlayDropdown';
+import style from '../Field.css';
 
 type Props = {
   children: React.Node,
